rustc: middle: move TraitItemKind from resolve to def.
